cord-cli: make required fields required
There are some request body properties that are required but only have one option, eg. `notification.type` must equal `url`, and `user.preferences.key` must equal `notification_channels`.

### Notification.type
<img width="862" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/1033e21b-5bc5-42a2-8e0d-97e63da6daaf">

### User.update-preferences.key
<img width="865" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/f0bc99be-e93c-49b0-8d57-8b446c107019">


Previously I was setting these options to be optional and pass a default of the only option. After speaking with @flooey it would be better to just make it required so there's no confusion if/when we add more possible values in the future.

While I was here I also noticed that we are passing a positional `id` param to the list all user command which doesn't require an `id`, so I updated that too:

### Specifying positional `id` param for each command:

##### before (see Positional description in `--help`):

##### after (no Positional description in `--help`):
<img width="698" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/cb1a75ce-cedc-47ae-9e78-1a786ef6cfad">



Test Plan: Run the commands

Reviewers: flooey, jwatzman, dmmiller

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6619

monorepo-commit: 27e3a97b155df69c398abb0dcbc25e7e2c084d69