cord-cli: create cord init command
Instead of clunkily having to add env variables to the console (which will then only live in that instance) we now have a `cord init` command! 🎈

This command will prompt the user to input the app id, app secret, and (if app management is needed then) customer id and customer secret. These variables will then be added to a `.cord` file that lives in the home directory.

##### Still to do!
I haven't tested this out on Windows yet. I'm hoping using `os.homedir()` is going to be friendly with Windows but I haven't had a chance to download a vm to test it yet 🤞 I will do this soon, but not in this PR - hope that's okay! 🙏

### Key features:
- if there is a `.cord` file already, the prompts will be populated with these default values
- there is a prompt after adding app id & app secret to ask if you will need app management
- there is still a chance to opt out of inputting app management credentials if you answer yes to the above prompt without loosing your previous answers by inputting an empty value


Test Plan: Run the command

Reviewers: dmmiller, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6632

monorepo-commit: 7e71a378152769d5320c12733afa4308811e464b