cord-cli: drain stdout before exiting
I noticed that `cord thread ls` in our testbed app (which has a ton of
threads) worked just fine, but `cord thread ls | jq` and `cord thread ls
| less` did not -- the output was truncated. After some poking around, I
found that the truncation happened after *exactly* 64k of output.

The issue is that node tells the OS not to block on a `write` syscall if
the buffer is full, and instead maintains its own internal buffer if
that starts happening. It will flush that buffer when it gets a chance,
but we need to wait to exit until that has happened.


Test Plan:
`cord thread ls | less` configured against our testbed no longer has
truncated output.

Reviewers: dmmiller, flooey, maylynn-ng

Reviewed By: dmmiller, maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/6795

monorepo-commit: 0a621c5e42f20edd5e3a5f577e6dec48a5392ea9