cord-cli: notify user if there's an update available
Linear ticket: E-2942

This checks the current version with the latest published one on npm and prints a message if there's an update available.

### To note:
I wasn't sure where to run this code. I was going to add it to the whole command, so it will always print. But when I saw that the call to npm slows down the command considerably, I decided against this and put it just in `cord`.

pros of this approach:
- doesn't affect the speed of the other calls

cons of this approach:
- have to specifically run `cord` to see the message

Any ideas what may be better?

Thought about having it run just on the first command, but then with a tool like this it's very hard to know what's the "first time" both theoretically and programmatically 🤷‍♀️

<img width="862" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/80ff505a-0c8f-4fc0-aa24-9fc691faae4c">
(👆 I hard coded the current version value to force show the message)

<br />
<br />


Test Plan: Run `cord`

Reviewers: jwatzman, dmmiller

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6793

monorepo-commit: a69468848c233ac7d487455b96f96364e653e128