CLI: add create thread command
Basically what the title says, we've got the new rest api for this so
adding it on the cli end as well.

Things to note:
- I'm using group-id as it is, without referencing organization-id as an
alias as this is a new endpoint on our REST API and doesn't use organizationID.
However, it's not consistent with our other commands so I'm happy to add
it in just for that.
- I have an error with ServerCreateThread's id prop no being optional, i
probably just need to publish the sdk package to get that update. i'll draft
this until I've done that


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/7082

monorepo-commit: e71caa26abe90eeb7ab3a6f22a00d739e6419ce6