Privacy: hacky REST API endpoints and CLI commands for create/delete permissions
This will make testing by hand a lot easier. The CLI commands will be
public, but everything is gated off server-side so they don't do
anything.

The API here, both for REST and the CLI, is likely to change, hence the
"experimental" prefix in the URL, which will mean that after we do
properly release this if someone tries with the old CLI they will get a
404. The CLI command also does not show up in the documentation.


Test Plan:
Used the CLI to create and then delete a rule. Looking at the raw row in
the local database, it was correct.

Reviewers: flooey, mhtien, yeo-yeo

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7279

monorepo-commit: 468c0b6b11bc9c6c174a3bb1f32ab0d3c7fef75c