cli: improvement to version checker
Our current version checker is very rudimentary and doesn't consider the case where we've updated the cli and so for a day it is ahead of the version the deployed monorepo knows.

The version in deployed monorepo is important because we are using that to check the most current version of the cli.

This improvement makes a proper check for the version rather than checking the equality of the string, so now it will not show that the cli is out of date if the cli is in fact ahead of monorepo.

ie.
<img width="382" alt="Untitled" src="https://github.com/getcord/monorepo/assets/62358728/46950b3a-cfde-49c9-9282-cde4aee959c1">



Test Plan: Try out the util `isLaterVersion` with a couple of examples

Reviewers: jwatzman, dmmiller, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/7369

monorepo-commit: 0b1b69e35ba58c10799d33f993d0dea97bdfa495