cord-cli: add application select to be able to update app level variables from a dropdown
This command allows you to select which app you would like to query within from a list that you can navigate through. Once you've selected one, it will update your `.cord` file with the appropriate `appID` and `appSecret`

The idea at first was to add a section to the terminal prompt that would highlight which app you would be in, though this would mean it would be present at all times, which probably wouldn't benefit our customers as it would us.

Instead, there is work coming up to add a `app which` command to see which app you're in, and also to add it to the `--help` page


Test Plan: Run the command

Reviewers: dmmiller, flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6831

monorepo-commit: b956f67aa8f3728344edf68da495c075327538ef