CLI: fix when curl returns a string
Dave found a bug where we'd get an error if the stdout return from curl is just a string, not json and so cannot be parsed. Thank you thank you @yeo-yeo for helping me spot this!

Wrapping this in a try catch sorts that out :phew:


Test Plan:
Run cord curl with a url that returns "Not found" or a 404, should return with a not found rather than error at json

Reviewers: yeo-yeo, dmmiller, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/8248

monorepo-commit: 8ede8c0a0f2de05ce45507ce6e42ca033da9377c