CLI: cord init to print out variables in bash format
Currently we are printing out the variables that have been added as a json object, though as Nimrod pointed out, not everyone reads the copy & may think that these values needed to be added to a .cord file instead of that they have been created for you.

So in case this does happen the log has been updated to have the right format for copy pasting into a .cord file.


Test Plan:
Run cord init, at the end the variables to be added should be logged in the format of `VARIABLE_NAME=VALUE`

Reviewers: dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/8266

monorepo-commit: ed72f1f52658a5c3b0a148ff286ab5ac6127fbb4